requirements.txt
streamlit==1.31.0
openai
python-dotenv
